Added a file selector for the csv file

+++ src/main/java/billboard/models/AreaChart.java
@@ -25,7 +25,7 @@ public class AreaChart {
         series.setName((String) settings.get("series"));
         
         for (Entry entry : entries) {
-            series.getData().add(new XYChart.Data(entry.getFirst(), Integer.parseInt(entry.getSecond())));
+            series.getData().add(new XYChart.Data(entry.getFirst(), Double.parseDouble(entry.getSecond())));
         }
         
         ac.getData().addAll(series);
