+++ src/main/java/billboard/models/BarGraph.java
@@ -26,7 +26,7 @@ public class BarGraph  {
         series.setName((String) settings.get("series"));
         
         for (Entry entry : entries) {
-            series.getData().add(new XYChart.Data(entry.getFirst(), Integer.parseInt(entry.getSecond())));
+            series.getData().add(new XYChart.Data(entry.getFirst(), Double.parseDouble(entry.getSecond())));
         }
         
         bc.getData().addAll(series);
